feat(InputField): migrate to Tailwind
